Backend/Mind/Creativity/requirements.txt
fastapi
uvicorn
pydantic
torch
diffusers
transformers
accelerate
pillow
